# contracts/core/openapi-case.yaml
openapi: 3.1.0
info:
  title: ResolveFlow case-service core API
  version: core-v1.2
  x-core-profile: core-v1.2
  x-core-schema-version: 2
  description: |
    Core-v1.2 case surface: the customer case lifecycle, the reviewer decisions, the
    operator reconciliation and the service-to-service endpoints the Agent needs.
    Target protocol of C00 (docs/core-contracts.md sections 3 and 4); it does not claim
    these operations are implemented yet.

    Narrower than the T02 compat document (contracts/openapi-case.yaml):
    - REFUND is the only executable action; reship is not implemented
      (docs/core-scope.md:9).
    - The refund operation vocabulary is the core five states
      (docs/domain-model.md:57), not the compat two-phase machine.
    - The three policy-management routes (/api/v1/policies/import, .../publish,
      .../revoke) are absent: policy is imported and validated by a controlled script
      (docs/core-contracts.md:53). The two read-only policy surfaces remain.
    - A case view returns its investigation trajectory (docs/core-scope.md:7), which the
      compat snapshot left to a separate timeline read.

    The public surface is authenticated with the demonstration user JWT; every
    /internal/v1/... operation requires the service token instead and is never exposed
    through the gateway.
servers:
  - url: https://api.resolveflow.local
    description: |
      Local compose deployment through the gateway; no public instance is published.
...
paths:
...
  /api/v1/orders:
    get:
      tags: [orders]
      operationId: listOrders
      summary: Order lines visible to the authenticated principal
      description: |
        A customer sees only their own lines, a merchant only its own; the scope comes
        from the token, so a crafted query cannot widen it (docs/core-contracts.md:15).
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: One page of order lines
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderLinePage'
              example:
                items:
                  - order_id: 00000000-0000-4000-8000-0000000000aa
                    line_id: '7001'
                    sku: SKU-RED-M
                    category: apparel
                    quantity: 2
                    line_paid_amount: 2599
                    currency: CNY
                    paid_at: '2026-09-10T08:15:00Z'
                    status: PAID
                    version: 3
                page:
                  limit: 20
                  next_cursor: null
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/orders/{order_id}:
    get:
      tags: [orders]
      operationId: readOrder
      summary: One authorised order and its lines
      description: |
        The customer's own order, or a merchant's order for a line it owns
        (docs/core-contracts.md:27). A line the principal cannot see is 404, not 403: the
        existence of another tenant's order is not disclosed.
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order lines the principal may see
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderLinePage'
              example:
                items:
                  - order_id: 00000000-0000-4000-8000-0000000000aa
                    line_id: '7001'
                    sku: SKU-RED-M
                    category: apparel
                    quantity: 2
                    line_paid_amount: 2599
                    currency: CNY
                    paid_at: '2026-09-10T08:15:00Z'
                    status: PAID
                    version: 3
                page:
                  limit: 20
                  next_cursor: null
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        User JWT from the demonstration login. It carries the principal; X-User, X-Role
        and X-Merchant headers are stripped at the gateway and never trusted here
        (docs/core-contracts.md:15).
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Internal aud/scope token for the /internal/v1/... surface, which is never exposed
        through the gateway.
  parameters:
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 200
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Unauthenticated:
      description: Missing or invalid user token (401)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHENTICATED
            message: the access token is missing or expired
            retryable: false
            trace_id: 0af7651916cd43dd8448eb211c80319c
...
    NotFound:
      description: Not visible under this principal (404)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: no case with that id for this principal
            retryable: false
            trace_id: 0af7651916cd43dd8448eb211c80319c
...
    ServiceUnavailable:
      description: A dependency is temporarily unavailable (503). A timeout is not a failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: SERVICE_UNAVAILABLE
            message: commerce-service did not answer within the reconciliation budget
            retryable: true
            trace_id: 0af7651916cd43dd8448eb211c80319c
  schemas:
    Uuid:
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
...
    Version:
      type: integer
      minimum: 1
      maximum: 9007199254740991
...
    AmountMinor:
      type: integer
      minimum: 0
      maximum: 1000000000
      description: CNY minor units, always an integer (docs/core-contracts.md:17).
...
    PageMeta:
      type: object
      additionalProperties: false
      required: [limit]
      properties:
        next_cursor:
          type: [string, 'null']
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message, retryable]
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 64
        message:
          type: string
          minLength: 1
          maxLength: 2000
        retryable:
          type: boolean
          description: |
            Carried explicitly so a caller never infers retryability from the status
            code. A network timeout is not a failed refund (docs/contracts.md:14).
        trace_id:
          type: [string, 'null']
          maxLength: 64
        details:
          type: [object, 'null']
          additionalProperties: true
    # The HTTP error body is the flat object docs/core-contracts.md:15 keeps unchanged
    # and com.resolveflow.shared.error.ApiError serialises.
    ErrorResponse:
      $ref: '#/components/schemas/ApiError'
...
    OrderLineSummary:
      type: object
      additionalProperties: false
      required: [order_id, line_id, sku, category, quantity, line_paid_amount, currency, paid_at, status, version]
      properties:
        order_id:
          $ref: '#/components/schemas/Uuid'
        line_id:
          type: string
          minLength: 1
          maxLength: 64
        sku:
          type: string
          minLength: 1
          maxLength: 64
        category:
          type: string
          minLength: 1
          maxLength: 64
        quantity:
          type: integer
          minimum: 1
          maximum: 1000
        line_paid_amount:
          $ref: '#/components/schemas/AmountMinor'
        currency:
          type: string
          const: CNY
        paid_at:
          type: string
          format: date-time
        status:
          type: string
          minLength: 1
          maxLength: 64
        version:
          $ref: '#/components/schemas/Version'
    OrderLinePage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderLineSummary'
        page:
          $ref: '#/components/schemas/PageMeta'
...
security:
  - bearerAuth: []